Emit DRED barrier breadcrumbs for barriers issued outside the flush path
