+++ trinityal/dx12/Tr2RenderContextDx12.cpp
@@ -1570,25 +1570,6 @@ void Tr2RenderContextAL::ResetDx12()
 	m_srgbWriteEnable = false;
 }
 
-namespace
-{
-// DRED breadcrumb contexts are only captured from PIX3-blob markers (Metadata=2,
-// WinPixEventRuntime encoding); legacy ANSI/unicode markers are ignored
-void SetDredMarker( ID3D12GraphicsCommandList* commandList, const char* text )
-{
-	constexpr UINT64 PIXEvent_SetMarker_NoArgs = 0x008;
-	UINT64 blob[64];
-	blob[0] = PIXEvent_SetMarker_NoArgs << 10; // timestamp 0, event type
-	blob[1] = 0xFF000000; // ARGB color
-	blob[2] = ( UINT64( 8 ) << 55 ) | ( UINT64( 1 ) << 54 ); // string info: copy chunk 8, isANSI
-	size_t lenBytes = strlen( text ) + 1;
-	size_t qwords = std::min( ( lenBytes + 7 ) / 8, size_t( 60 ) );
-	memset( &blob[3], 0, qwords * 8 );
-	memcpy( &blob[3], text, std::min( lenBytes, qwords * 8 - 1 ) );
-	commandList->SetMarker( 2, blob, UINT( ( 3 + qwords ) * 8 ) );
-}
-}
-
 void Tr2RenderContextAL::AddGpuMarker( const char* marker )
 {
 	m_ownerDevice->GetMarkerBuffer().PutMarker( m_commandList2, marker );
@@ -1599,7 +1580,7 @@ void Tr2RenderContextAL::AddGpuMarker( const char* marker )
 	}
 	if( g_dredBreadcrumbsEnabled )
 	{
-		SetDredMarker( m_commandList, marker );
+		TrinityALImpl::SetDredMarker( m_commandList, marker );
 	}
 }
 
@@ -1745,110 +1726,11 @@ void Tr2RenderContextAL::ResourceBarrierDx12( const D3D12_RESOURCE_BARRIER& barr
 	ResourceBarrierDx12( 1, &barrier );
 }
 
-namespace
-{
-// snprintf returns the untruncated length; clamp so pos never passes the terminator
-size_t AdvanceFormatPos( size_t pos, int written, size_t size )
-{
-	return written < 0 || pos + size_t( written ) >= size ? size - 1 : pos + size_t( written );
-}
-
-const char* FormatResourceStates( char* buf, size_t size, D3D12_RESOURCE_STATES states )
-{
-	if( states == D3D12_RESOURCE_STATE_COMMON )
-	{
-		return "COMMON";
-	}
-	if( states == D3D12_RESOURCE_STATE_GENERIC_READ )
-	{
-		return "GENERIC_READ";
-	}
-	static const struct
-	{
-		D3D12_RESOURCE_STATES bit;
-		const char* name;
-	} s_stateNames[] = {
-		{ D3D12_RESOURCE_STATE_VERTEX_AND_CONSTANT_BUFFER, "VB_CB" },
-		{ D3D12_RESOURCE_STATE_INDEX_BUFFER, "IB" },
-		{ D3D12_RESOURCE_STATE_RENDER_TARGET, "RT" },
-		{ D3D12_RESOURCE_STATE_UNORDERED_ACCESS, "UAV" },
-		{ D3D12_RESOURCE_STATE_DEPTH_WRITE, "DEPTH_W" },
-		{ D3D12_RESOURCE_STATE_DEPTH_READ, "DEPTH_R" },
-		{ D3D12_RESOURCE_STATE_NON_PIXEL_SHADER_RESOURCE, "SRV_NONPX" },
-		{ D3D12_RESOURCE_STATE_PIXEL_SHADER_RESOURCE, "SRV_PX" },
-		{ D3D12_RESOURCE_STATE_STREAM_OUT, "STREAM_OUT" },
-		{ D3D12_RESOURCE_STATE_INDIRECT_ARGUMENT, "INDIRECT" },
-		{ D3D12_RESOURCE_STATE_COPY_DEST, "COPY_DST" },
-		{ D3D12_RESOURCE_STATE_COPY_SOURCE, "COPY_SRC" },
-		{ D3D12_RESOURCE_STATE_RESOLVE_DEST, "RESOLVE_DST" },
-		{ D3D12_RESOURCE_STATE_RESOLVE_SOURCE, "RESOLVE_SRC" },
-		{ D3D12_RESOURCE_STATE_RAYTRACING_ACCELERATION_STRUCTURE, "RTAS" },
-	};
-	size_t pos = 0;
-	UINT remaining = UINT( states );
-	for( const auto& state : s_stateNames )
-	{
-		if( ( remaining & UINT( state.bit ) ) == UINT( state.bit ) )
-		{
-			pos = AdvanceFormatPos( pos, snprintf( buf + pos, size - pos, "%s%s", pos ? "|" : "", state.name ), size );
-			remaining &= ~UINT( state.bit );
-		}
-	}
-	if( remaining )
-	{
-		snprintf( buf + pos, size - pos, "%s0x%x", pos ? "|" : "", remaining );
-	}
-	return buf;
-}
-
-// Recorded immediately before each ResourceBarrier so the DRED breadcrumb context
-// identifies which resources/states the otherwise anonymous RESOURCEBARRIER op contains
-void EmitBarrierBreadcrumb( ID3D12GraphicsCommandList* commandList, const D3D12_RESOURCE_BARRIER* barriers, size_t count )
-{
-	if( !g_dredBreadcrumbsEnabled )
-	{
-		return;
-	}
-	char buf[512];
-	size_t pos = AdvanceFormatPos( 0, snprintf( buf, sizeof( buf ), "[Barrier]" ), sizeof( buf ) );
-	for( size_t i = 0; i < count && pos < sizeof( buf ) - 1; ++i )
-	{
-		const auto& barrier = barriers[i];
-		ID3D12Resource* resource = barrier.Type == D3D12_RESOURCE_BARRIER_TYPE_UAV ? barrier.UAV.pResource : barrier.Type == D3D12_RESOURCE_BARRIER_TYPE_ALIASING ? barrier.Aliasing.pResourceAfter :
-																																									barrier.Transition.pResource;
-		char name[128];
-		UINT nameSize = sizeof( name ) - 1;
-		if( !resource || FAILED( resource->GetPrivateData( WKPDID_D3DDebugObjectName, &nameSize, name ) ) || nameSize >= sizeof( name ) )
-		{
-			nameSize = UINT( snprintf( name, sizeof( name ), "%p", resource ) );
-		}
-		name[nameSize] = 0;
-		if( barrier.Type == D3D12_RESOURCE_BARRIER_TYPE_TRANSITION )
-		{
-			char before[96], after[96];
-			const char* beforeStates = FormatResourceStates( before, sizeof( before ), barrier.Transition.StateBefore );
-			const char* afterStates = FormatResourceStates( after, sizeof( after ), barrier.Transition.StateAfter );
-			pos = AdvanceFormatPos( pos, snprintf( buf + pos, sizeof( buf ) - pos, " %s(%s->%s)", name, beforeStates, afterStates ), sizeof( buf ) );
-		}
-		else if( barrier.Type == D3D12_RESOURCE_BARRIER_TYPE_UAV )
-		{
-			pos = AdvanceFormatPos( pos, snprintf( buf + pos, sizeof( buf ) - pos, " UAV(%s)", name ), sizeof( buf ) );
-		}
-		else
-		{
-			pos = AdvanceFormatPos( pos, snprintf( buf + pos, sizeof( buf ) - pos, " Alias(%s)", name ), sizeof( buf ) );
-		}
-	}
-	SetDredMarker( commandList, buf );
-}
-}
-
 void Tr2RenderContextAL::FlushBarriersDx12()
 {
 	if( !m_barriers.empty() )
 	{
-		EmitBarrierBreadcrumb( m_commandList, m_barriers.data(), m_barriers.size() );
-		m_commandList->ResourceBarrier( UINT( m_barriers.size() ), m_barriers.data() );
+		TrinityALImpl::ResourceBarrier( m_commandList, UINT( m_barriers.size() ), m_barriers.data() );
 		m_barriers.clear();
 	}
 }
@@ -1895,8 +1777,7 @@ void Tr2RenderContextAL::FlushBarriersDx12( size_t count, ID3D12Resource** resou
 		}
 		if( barrierCount )
 		{
-			EmitBarrierBreadcrumb( m_commandList, barriers, barrierCount );
-			m_commandList->ResourceBarrier( UINT( barrierCount ), barriers );
+			TrinityALImpl::ResourceBarrier( m_commandList, UINT( barrierCount ), barriers );
 		}
 	}
 	else
@@ -1922,8 +1803,7 @@ void Tr2RenderContextAL::FlushBarriersDx12( size_t count, ID3D12Resource** resou
 		}
 		if( !barriers.empty() )
 		{
-			EmitBarrierBreadcrumb( m_commandList, barriers.data(), barriers.size() );
-			m_commandList->ResourceBarrier( UINT( barriers.size() ), barriers.data() );
+			TrinityALImpl::ResourceBarrier( m_commandList, UINT( barriers.size() ), barriers.data() );
 		}
 	}
 }
